requirements.txt
anyio==4.4.0
boto3==1.35.15
botocore==1.35.15
certifi==2024.8.30
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.5
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
urllib3==2.2.2
 